relocate skill rows and rename SkillEntries
